+++ src/Modules/users/users.service.ts
@@ -1,7 +1,7 @@
 import {
   BadRequestException,
   Injectable,
-  InternalServerErrorException,
+  NotFoundException
 } from '@nestjs/common';
 import User from '../../entities/users.entity';
 import { InjectRepository } from '@nestjs/typeorm';
@@ -17,7 +17,6 @@ export class UsersService {
     page: number,
     limit: number,
   ): Promise<Omit<User, 'isAdmin' | 'password'>[]> {
-    try {
       const start = (page - 1) * limit;
       const end = start + limit;
       const users: User[] = await this.userRepository.find();
@@ -27,38 +26,25 @@ export class UsersService {
       return usersList.map(
         ({ password, isAdmin, ...userNoPassword }) => userNoPassword,
       );
-    } catch (error) {
-      throw new InternalServerErrorException(
-        'Error al intentar traer los Usuarios',
-      );
-    }
   }
 
   async deleteUsers(id: string): Promise<Object> {
-    try {
       const user: User = await this.userRepository.findOneBy({ id });
+      if(user.isAdmin==="admin") throw new NotFoundException("No se puede eliminar un administrador")
       if (!user)
         throw new BadRequestException(`El usuario con id: ${id} no existe`);
       await this.userRepository.delete(id);
       return { message: 'El usuario ha sido eliminado' };
-    } catch (error) {
-      throw new InternalServerErrorException(
-        'Error al intentar Eliminar el usuario',
-      );
     }
-  }
 
-  async updateUsers(id: string, user: Partial<UserDto>): Promise<Object> {
-    try {
+  async updateUsers(id: string, user: UserDto): Promise<Object> {
       const usuario: User = await this.userRepository.findOneBy({ id });
       if (!usuario)
         throw new BadRequestException(`El usuario con id: ${id} no existe`);
       await this.userRepository.update(id, user);
       return { message: 'Usuario Actualizado Correctamente' };
-    } catch (error) {
-      throw new InternalServerErrorException('Error al actualizar el Usuario');
     }
-  }
+  
 
   async getUserById(id: string): Promise<Omit<User, 'password' | 'isAdmin'>> {
       const users = await this.userRepository.findOne({
